Replace 32-bit cast ratchet with a per-site #[expect] gate (#72)
The `cast-ratchet-32bit` job capped the *total count* of i686
narrowing-cast warnings at a baseline of 268. As raised in #72 this had
two flaws: the baseline was never ratcheted down, so fixes weren't
locked in, and a count cap is fungible — a genuinely new cast passes CI
as long as an unrelated one is removed in the same PR.

Replace it with `cast-deny-32bit`, which hard-denies
clippy::cast_possible_truncation, clippy::cast_possible_wrap, and
unfulfilled_lint_expectations on the pinned i686 toolchain. Every
narrowing cast in the library is now either:

- a checked crate::convert conversion — the genuine file-derived
  index/size hazards: the chunked-data reader and the MATLAB matrix
  reader now return an error instead of silently truncating when a
  dimension or element count exceeds the platform's pointer width; or
- annotated at its site with #[expect(..., reason = "...")] stating the
  bound that makes it safe.

The gate is non-fungible (a new unannotated cast fails on its own) and
self-cleaning (a stale #[expect] left after its cast is removed also
fails). Updates the convert.rs guidance and CHANGELOG.

No public API changes. Verified: i686 deny gate clean, check-32bit (both
configs), host clippy --all-targets -D warnings, fmt, no-std, wasm, and
the full test suite (479 lib + integration) all pass.

+++ src/mat/userblock.rs
@@ -18,7 +18,7 @@ pub const DEFAULT_DESCRIPTION: &str = "MATLAB 7.3 MAT-file, Platform: hdf5-pure 
 /// untouched (zeros from `FileBuilder::with_userblock(512)`).
 pub fn write_header(file_bytes: &mut [u8], description: &str) {
     assert!(
-        file_bytes.len() >= USERBLOCK_SIZE as usize,
+        file_bytes.len() as u64 >= USERBLOCK_SIZE,
         "userblock requires at least 512 bytes, got {}",
         file_bytes.len()
     );
@@ -52,7 +52,7 @@ pub fn write_header(file_bytes: &mut [u8], description: &str) {
 /// Verify the bytes look like a MATLAB v7.3 userblock. Returns `Ok(())` on
 /// success.
 pub fn verify_header(file_bytes: &[u8]) -> Result<(), &'static str> {
-    if file_bytes.len() < USERBLOCK_SIZE as usize {
+    if (file_bytes.len() as u64) < USERBLOCK_SIZE {
         return Err("file too short for MAT v7.3 userblock");
     }
     if !file_bytes[..6].eq_ignore_ascii_case(b"MATLAB") {
